+++ src/components/Header/Weather/Weather.module.css
@@ -1,7 +1,7 @@
 .weatherContainer {
   position: absolute;
   top: 0.8rem;
-  right: 1rem;
+  right: -1rem;
   color: #fff;
   line-height: normal;
 }
@@ -30,4 +30,10 @@
 .weatherModalImg {
   width: 100%;
   margin-bottom: 10px;
+}
+
+@media only screen and (max-width: 768px) {
+  .content{
+    display: none;
+  }
 }
