+++ src/components/Header/Weather/WeatherComponent.jsx
@@ -79,6 +79,7 @@ const WeatherComponent = () => {
       <Col
         onClick={showModal}
         style={{ cursor: 'pointer' }}
+        className={style.content}
       >
         <Row justify="center">
           <Text className={style.weatherTitle}>
