+++ src/components/Main/Exercise/ExerciseSchedule/ExerciseSchedule.module.css
@@ -16,6 +16,12 @@
   }
 }
 
+@media only screen and (max-width: 1230px) {
+  .content {
+    margin: 3px;
+  }
+}
+
 .content::-webkit-scrollbar{
   width: 1px;
 }
@@ -27,6 +33,10 @@
   }
 }
 
+.add_form_content {
+  width: 14.9rem;
+}
+
 .selected {
   margin: 0 auto;
   border: 1.5px solid darkblue;
@@ -35,10 +45,6 @@
 @media only screen and (max-width: 805px) {
   .selected {
     display: block;
-  }
-
-  .add_wrapper {
-    margin-top: 10px;
   }
 }
